+++ Objects/ItemInformation.cs
@@ -4,10 +4,13 @@
     {
         public int maxValue {  get; set; }
         public int minValue {  get; set; }
+        public bool conductive { get; set; }
 
-        public ItemInformation(int min, int max) {
+        public ItemInformation(int min, int max, bool conductive)
+        {
             maxValue = max;
             minValue = min;
+            this.conductive = conductive;
         }
     }
 }